consolidate mdbase into one native CLI
