+++ deploy/docker/mdbase-rs-revision
@@ -1,1 +1,1 @@
-b7072ae4e5191c00956ed64913684af2c5ad9ccd
+81ee2e7bfde665251316740ee92cf4fdb66c6196
