+++ docs/architecture.md
@@ -38,7 +38,7 @@ never identities and never leave the connector. If a copied folder presents
 the same ID while the original remains registered, the connector rejects the
 copy and asks the user to establish a distinct identity instead of silently
 aliasing two authorities. The desktop's explicit **Register copy** action, or
-`mdbase-connect collection add-copy PATH`, writes a new random ID only after
+`mdbase connect collection add-copy PATH`, writes a new random ID only after
 the connector has proved that the selected path is a copy of a different,
 still-registered folder. It refuses the registered original, a moved folder,
 and a collection that can be registered normally.
