+++ docs/releasing.md
@@ -1,7 +1,7 @@
 # Beta release checklist
 
 mdbase connect desktop bundles contain the Electron client and the matching
-Rust `mdbase-connect` daemon/CLI. A release is one tested unit; mixing desktop
+Rust `mdbase` daemon/CLI. A release is one tested unit; mixing desktop
 and daemon versions is unsupported.
 
 Until the stable `0.1.0` contract is ready, releases use
@@ -31,7 +31,7 @@ pnpm --filter @mdbase/connect-desktop package
 ```
 
 The package command compiles the release daemon/CLI, creates the platform
-Electron bundle, and verifies that `app.asar` and the `mdbase-connect`
+Electron bundle, and verifies that `app.asar` and the `mdbase`
 executable are both present.
 Run the packaged application once with a fresh user-data directory and complete
 pairing, collection registration, encrypted application authorization, one write,
