+++ package.json
@@ -14,11 +14,11 @@
     "dev:desktop:fresh": "node scripts/isolated-desktop.mjs --fresh",
     "e2e": "pnpm build && cargo build --workspace && node scripts/e2e.mjs",
     "e2e:container": "node scripts/container-e2e.mjs",
-    "e2e:desktop:container": "pnpm --filter @mdbase/connect-desktop build:standalone && cargo build -p mdbase-connect && node apps/desktop/scripts/docker-server-e2e.mjs",
+    "e2e:desktop:container": "pnpm --filter @mdbase/connect-desktop build:standalone && cargo build -p mdbase-cli && node apps/desktop/scripts/docker-server-e2e.mjs",
     "e2e:ecosystem": "pnpm e2e:container && pnpm e2e:desktop:container",
     "e2e:relay": "pnpm build && node scripts/relay-e2e.mjs",
     "e2e:sync": "pnpm build && node scripts/sync-e2e.mjs",
-    "e2e:provider": "pnpm build && cargo build -p mdbase-connect-hosted-provider -p mdbase-connect && node scripts/hosted-provider-e2e.mjs",
+    "e2e:provider": "pnpm build && cargo build -p mdbase-connect-hosted-provider -p mdbase-cli && node scripts/hosted-provider-e2e.mjs",
     "e2e:authority:stress": "MDBASE_CONNECT_PROVIDER_E2E_PROMOTION_COUNT=2500 pnpm e2e:provider",
     "e2e:provider:stress": "MDBASE_CONNECT_PROVIDER_E2E_BULK_COUNT=10000 pnpm e2e:provider",
     "profile:mirror": "node --expose-gc scripts/mirror-profile.mjs",
@@ -28,9 +28,9 @@
     "check:mirror:mobile": "pnpm --filter @mdbase/connect-sync check:mobile",
     "package:audit": "pnpm build && node scripts/package-audit.mjs",
     "package:consumer": "node scripts/pack-consumer-sdk.mjs",
-    "profile:connect": "cargo run --release -p mdbase-connect-core --bin connect-profile --",
-    "profile:rs": "cargo run --release --manifest-path ../mdbase-rs/Cargo.toml --bin mdb-profile -- --scenario queries",
-    "profile:views:rs": "cargo run --release --manifest-path ../mdbase-rs/Cargo.toml --bin mdb-profile -- --scenario views",
+    "profile:connect": "cargo run --release -p mdbase-cli -- profile connect",
+    "profile:rs": "cargo run --release -p mdbase-cli -- profile engine --scenario queries",
+    "profile:views:rs": "cargo run --release -p mdbase-cli -- profile engine --scenario views",
     "test": "node --test scripts/lib/*.test.mjs && pnpm --filter @mdbase/connect-desktop test && pnpm --filter '!@mdbase/connect-desktop' -r test",
     "typecheck": "pnpm -r typecheck",
     "version:check": "node scripts/check-release-version.mjs"
