+++ scripts/verify-windows-packages.ps1
@@ -47,7 +47,7 @@ $connector = Get-SinglePackageFile `
   -Files @(
     Get-ChildItem (Join-Path $packageDirectory.FullName "resources") `
       -File `
-      -Filter "mdbase-connect.exe"
+      -Filter "mdbase.exe"
   )
 $installer = Get-SinglePackageFile `
   -Description "Squirrel installer" `
@@ -150,7 +150,7 @@ if ([string]$properties.PublisherDisplayName -ne $PublisherDisplayName) {
 if ($applicationNode.Executable -ne "app\mdbase-connect.exe") {
   throw "Unexpected Store package executable: $($applicationNode.Executable)"
 }
-$storeConnector = Join-Path $unpacked "app\resources\mdbase-connect.exe"
+$storeConnector = Join-Path $unpacked "app\resources\mdbase.exe"
 if (-not (Test-Path $storeConnector -PathType Leaf)) {
   throw "Microsoft Store package does not contain the connector CLI"
 }
